requirements.txt
Django==3.2.17
djangorestframework==3.14.0